test: Fuzz execution of a single instruction with afl-fuzz (#37)
Replaces stack, program and heap with mocks that just contain one value
and crash if multiple addresses are accessed. Fuzzes the execution of a
single instruction.

The default settings of libFuzzer don't work for me and it doesn't tell
what is going wrong, so I wrote this for afl-fuzz instead.

Afl explores lots of instructions in a matter of seconds, though the
crashes it finds are not real ones yet because the test is still missing
a state validity check.

The validity check checks that fat pointers are well-formed.
The test is not run if the initial state is invalid. The test panics if the state is no longer valid after running an
instruction.