store program counter inside vm
This way there is no need for the integration to juggle it when
the VM is suspended. It also enables tracers to manipulate it.

This also changes the pc storage format of previous frames and
near calls from u16 to *const Instruction, which increases memory
use and may hurt performance, so may need to be changed.